+++ aidial_adapter_openai/audio_api/speech/configuration.py
